requirements.txt
asttokens==2.2.1
astunparse==1.6.3
backcall==0.2.0
backports.functools-lru-cache==1.6.4
contourpy==1.0.7
cycler==0.11.0
debugpy==1.5.1
decorator==5.1.1
dill==0.3.6
dominate==2.7.0
entrypoints==0.4
executing==1.2.0
fonttools==4.39.3
imageio==2.29.0
importlib-resources==5.12.0
ipdb==0.13.13
ipykernel==6.15.0
ipython==8.4.0
jedi==0.18.2
jittor==1.3.7.14
jupyter-client==7.0.6
jupyter_core==5.3.0
kiwisolver==1.4.4
lazy_loader==0.2
matplotlib==3.7.1
matplotlib-inline==0.1.6
nest-asyncio==1.5.6
networkx==3.1
numpy==1.24.3
opencv-python==4.7.0.72
packaging==23.1
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.5.0
pip==23.0.1
platformdirs==3.5.0
prompt-toolkit==3.0.38
psutil==5.9.0
ptyprocess==0.7.0
pure-eval==0.2.2
Pygments==2.15.1
pyparsing==3.0.9
python-dateutil==2.8.2
PyWavelets==1.4.1
pyzmq==19.0.2
scikit-image==0.20.0
scipy==1.9.1
setuptools==66.0.0
six==1.16.0
stack-data==0.6.2
tifffile==2023.4.12
tomli==2.0.1
tornado==6.1
tqdm==4.65.0
traitlets==5.9.0
typing_extensions==4.5.0
wcwidth==0.2.6
wheel==0.38.4
zipp==3.15.0
